+++ tests/test_utils.py
@@ -1,5 +1,6 @@
 from scim2_models import EnterpriseUser
 from scim2_models import Group
+from scim2_models import ServiceProviderConfig
 from scim2_models import User
 
 from scim2_client import SCIMClient
@@ -10,3 +11,6 @@ def test_guess_resource_endpoint():
     assert client.resource_endpoint(Group) == "/Groups"
     assert client.resource_endpoint(User) == "/Users"
     assert client.resource_endpoint(User[EnterpriseUser]) == "/Users"
+
+    # This one is special as it does not take an ending 's'
+    assert client.resource_endpoint(ServiceProviderConfig) == "/ServiceProviderConfig"